requirements.txt
django==4.2.7
djangorestframework==3.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0